refactor(core): unisce csv_quick_shape in parquet.py, mantiene shim backward compat
