--- toolkit/core/manifest.py
@@ -1,6 +1,0 @@
-"""Deprecated — ``manifest.json`` has been removed in favor of ``metadata.json``.
-
-Previously this module contained ``write_raw_manifest()`` and
-``read_raw_manifest()`` which wrote/read ``manifest.json``.
-All data is now stored in ``metadata.json`` via ``toolkit.core.metadata``.
-"""
